Fix: fix delete logic for db optimization
유저 삭제 -> 카드의 유저 리스트 삭제후 유저 리스트가 없으면 카드도 삭제.

+++ src/services/user-service.ts
@@ -1,4 +1,4 @@
-import { userClient } from '@/client';
+import { cardClient, userClient } from '@/client';
 import { CardDto } from '@/domain/card';
 import { User } from '@/domain/user';
 import { UserDto } from '@/domain/user/user-dto';
@@ -25,10 +25,38 @@ export const userService = {
   ): Promise<boolean> => {
     return await userClient.update({ id, nickname, phone });
   },
-  updateCardList: async (id: string, cardList: string[]): Promise<boolean> => {
+  updateCardIdList: async (
+    id: string,
+    cardList: string[],
+  ): Promise<boolean> => {
     return await userClient.update({ id, cardIdList: cardList });
   },
   delete: async (id: string): Promise<boolean> => {
+    const user = await userClient.getById(id);
+    if (!user) return false;
+
+    user.cardIdList?.forEach(async cardId => {
+      const card = (await cardClient.getById(cardId)) as CardDto;
+      card.ownerList = card.ownerList.filter(userId => userId !== id);
+      if (card.ownerList.length === 0) await cardClient.deleteById(cardId);
+      else await cardClient.update(card);
+    });
+
     return await userClient.deleteById(id);
+  },
+  deleteCard: async (userId: string, cardId: string): Promise<boolean> => {
+    const user = await userClient.getById(userId);
+    const card = await cardClient.getById(cardId);
+    if (!user || !card) return false;
+
+    const newCardIdList = user.cardIdList.filter(id => id !== cardId);
+    user.cardIdList = newCardIdList;
+    userClient.update(user);
+
+    const newCardOwnerList = card.ownerList.filter(id => id !== userId);
+    if (newCardOwnerList.length === 0) return cardClient.deleteById(cardId);
+
+    card.ownerList = newCardOwnerList;
+    return cardClient.update(card);
   },
 };
